# docs/swagger.yaml
definitions:
  big.Int:
    type: object
...
  db.GetAllTourRow:
    properties:
      anh_chinh:
        type: string
      avg_rating:
        type: number
      danh_muc_id:
        type: integer
      danh_muc_ten:
        type: string
      diem_den: {}
      don_vi_tien_te:
        type: string
      gia_nguoi_lon:
        $ref: '#/definitions/pgtype.Numeric'
      gia_sau_giam_nguoi_lon: {}
      gia_sau_giam_tre_em: {}
      gia_tre_em:
        $ref: '#/definitions/pgtype.Numeric'
      giam_gia_phan_tram: {}
      id:
        type: integer
      mo_ta:
        type: string
      next_departure_date: {}
      nha_cung_cap_ten:
        type: string
      noi_bat:
        type: boolean
      so_dem:
        type: integer
      so_ngay:
        type: integer
      tieu_de:
        type: string
      total_reviews:
        type: integer
      trang_thai:
        type: string
    type: object
...
  gin.H:
    additionalProperties: {}
    type: object
...
  models.CreateTourFullRequest:
    properties:
      cau_hinh_nhom_tours:
        description: Cấu hình nhóm (optional)
        properties:
          so_lon_nhat:
            type: integer
          so_nho_nhat:
            type: integer
        type: object
      danh_muc_id:
        type: integer
      dia_diem_tours:
        description: Điểm đến
        items:
          properties:
            diem_den_id:
              type: integer
            thu_tu_tham_quan:
              type: integer
          required:
          - diem_den_id
          type: object
        type: array
      don_vi_tien_te:
        type: string
      gia_nguoi_lon:
        type: number
      gia_tre_em:
        type: number
      hinh_anh_tours:
        description: Ảnh tour
        items:
          properties:
            la_anh_chinh:
              type: boolean
            link:
              type: string
            mo_ta_alt:
              type: string
            thu_tu_hien_thi:
              type: integer
          required:
          - link
          type: object
        type: array
      lich_khoi_hanh_tours:
        description: Lịch khởi hành (optional)
        items:
          properties:
            ghi_chu:
              description: Optional notes
              type: string
            ngay_ket_thuc:
              description: 'Format: "YYYY-MM-DD"'
              type: string
            ngay_khoi_hanh:
              description: 'Format: "YYYY-MM-DD"'
              type: string
            suc_chua:
              description: Capacity
              minimum: 1
              type: integer
            trang_thai:
              description: 'Optional: len_lich, xac_nhan, huy, hoan_thanh'
              type: string
          required:
          - ngay_ket_thuc
          - ngay_khoi_hanh
          - suc_chua
          type: object
        type: array
      lich_trinh_tours:
        description: Lịch trình với hoạt động
        items:
          properties:
            dia_diem:
              type: string
            gio_bat_dau:
              description: 'Format: "HH:MM:SS"'
              type: string
            gio_ket_thuc:
              description: 'Format: "HH:MM:SS"'
              type: string
            hoat_dong_lich_trinh_tours:
              description: Hoạt động trong ngày
              items:
                properties:
                  gio_bat_dau:
                    description: 'Format: "HH:MM:SS"'
                    type: string
                  gio_ket_thuc:
                    description: 'Format: "HH:MM:SS"'
                    type: string
                  mo_ta:
                    type: string
                  ten:
                    type: string
                  thu_tu:
                    type: integer
                required:
                - ten
                type: object
              type: array
            mo_ta:
              type: string
            ngay_thu:
              type: integer
            thong_tin_luu_tru:
              type: string
            tieu_de:
              type: string
          required:
          - ngay_thu
          - tieu_de
          type: object
        type: array
      mo_ta:
        type: string
      nha_cung_cap_id:
        type: integer
      noi_bat:
        type: boolean
      so_dem:
        minimum: 0
        type: integer
      so_ngay:
        minimum: 1
        type: integer
      tieu_de:
        description: Thông tin tour cơ bản
        type: string
      trang_thai:
        type: string
    required:
    - gia_nguoi_lon
    - gia_tre_em
    - nha_cung_cap_id
    - so_ngay
    - tieu_de
    type: object
...
  pgtype.InfinityModifier:
    enum:
    - 1
    - 0
    - -1
    format: int32
    type: integer
    x-enum-varnames:
    - Infinity
    - Finite
    - NegativeInfinity
  pgtype.Numeric:
    properties:
      exp:
        format: int32
        type: integer
      infinityModifier:
        $ref: '#/definitions/pgtype.InfinityModifier'
      int:
        $ref: '#/definitions/big.Int'
      naN:
        type: boolean
      valid:
        type: boolean
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Travia API for travel management
  termsOfService: http://swagger.io/terms/
  title: Travia API
  version: "1.0"
paths:
...
  /tour:
    post:
      consumes:
      - application/json
      description: Tạo tour bao gồm ảnh, điểm đến, lịch trình và hoạt động trong 1
        transaction. Đảm bảo tính toàn vẹn dữ liệu (all or nothing)
      parameters:
      - description: Tour data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateTourFullRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Tạo tour với đầy đủ thông tin (1 transaction)
      tags:
      - tour
  /tour/:
    get:
      consumes:
      - application/json
      description: Lấy danh sách tour
      parameters:
      - default: 10
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 1
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Thành công
          schema:
            $ref: '#/definitions/db.GetAllTourRow'
        "500":
          description: Lỗi server
          schema:
            $ref: '#/definitions/gin.H'
      summary: Lấy danh sách tour
      tags:
      - tour
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"